Struct google_cloudchannel1::api::AccountMethods

pub struct AccountMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on account resources. It is not used directly, but through the Cloudchannel hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_cloudchannel1 as cloudchannel1;
 
use std::default::Default;
use oauth2;
use cloudchannel1::Cloudchannel;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = Cloudchannel::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `channel_partner_links_create(...)`, `channel_partner_links_customers_create(...)`, `channel_partner_links_customers_delete(...)`, `channel_partner_links_customers_get(...)`, `channel_partner_links_customers_list(...)`, `channel_partner_links_customers_patch(...)`, `channel_partner_links_get(...)`, `channel_partner_links_list(...)`, `channel_partner_links_patch(...)`, `check_cloud_identity_accounts_exist(...)`, `customers_create(...)`, `customers_delete(...)`, `customers_entitlements_activate(...)`, `customers_entitlements_cancel(...)`, `customers_entitlements_change_offer(...)`, `customers_entitlements_change_parameters(...)`, `customers_entitlements_change_renewal_settings(...)`, `customers_entitlements_create(...)`, `customers_entitlements_get(...)`, `customers_entitlements_list(...)`, `customers_entitlements_start_paid_service(...)`, `customers_entitlements_suspend(...)`, `customers_get(...)`, `customers_list(...)`, `customers_list_purchasable_offers(...)`, `customers_list_purchasable_skus(...)`, `customers_patch(...)`, `customers_provision_cloud_identity(...)`, `customers_transfer_entitlements(...)`, `customers_transfer_entitlements_to_google(...)`, `list_subscribers(...)`, `list_transferable_offers(...)`, `list_transferable_skus(...)`, `offers_list(...)`, `register(...)` and `unregister(...)`
// to build up your call.
let rb = hub.accounts();

Implementations

impl<'a, C> AccountMethods<'a, C>

pub fn channel_partner_links_customers_create(
&self,
request: GoogleCloudChannelV1Customer,
parent: &str
) -> AccountChannelPartnerLinkCustomerCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new Customer resource under the reseller or distributor account. Possible error codes: * PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request. * INVALID_ARGUMENT: * Required request parameters are missing or invalid. * Domain field value doesn't match the primary email domain. Return value: The newly created Customer resource.

Arguments

request - No description provided.

parent - Required. The resource name of reseller account in which to create the customer. Parent uses the format: accounts/{account_id}

pub fn channel_partner_links_customers_delete(
&self,
name: &str
) -> AccountChannelPartnerLinkCustomerDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes the given Customer permanently and irreversibly. Possible error codes: * PERMISSION_DENIED: The account making the request does not own this customer. * INVALID_ARGUMENT: Required request parameters are missing or invalid. * FAILED_PRECONDITION: The customer has existing entitlements. * NOT_FOUND: No Customer resource found for the name in the request.

Arguments

name - Required. The resource name of the customer to delete.

pub fn channel_partner_links_customers_get(
&self,
name: &str
) -> AccountChannelPartnerLinkCustomerGetCall<'a, C>

Create a builder to help you perform the following task:

Returns a requested Customer resource. Possible error codes: * PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request. * INVALID_ARGUMENT: Required request parameters are missing or invalid. * NOT_FOUND: The customer resource doesn't exist. Usually the result of an invalid name parameter. Return value: The Customer resource.

Arguments

name - Required. The resource name of the customer to retrieve. Name uses the format: accounts/{account_id}/customers/{customer_id}

pub fn channel_partner_links_customers_list(
&self,
parent: &str
) -> AccountChannelPartnerLinkCustomerListCall<'a, C>

Create a builder to help you perform the following task:

List Customers. Possible error codes: * PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request. * INVALID_ARGUMENT: Required request parameters are missing or invalid. Return value: List of Customers, or an empty list if there are no customers.

Arguments

parent - Required. The resource name of the reseller account to list customers from. Parent uses the format: accounts/{account_id}.

pub fn channel_partner_links_customers_patch(
&self,
request: GoogleCloudChannelV1Customer,
name: &str
) -> AccountChannelPartnerLinkCustomerPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates an existing Customer resource for the reseller or distributor. Possible error codes: * PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request. * INVALID_ARGUMENT: Required request parameters are missing or invalid. * NOT_FOUND: No Customer resource found for the name in the request. Return value: The updated Customer resource.

Arguments

request - No description provided.

name - Output only. Resource name of the customer. Format: accounts/{account_id}/customers/{customer_id}

pub fn channel_partner_links_create(
&self,
request: GoogleCloudChannelV1ChannelPartnerLink,
parent: &str
) -> AccountChannelPartnerLinkCreateCall<'a, C>

Create a builder to help you perform the following task:

Initiates a channel partner link between a distributor and a reseller, or between resellers in an n-tier reseller channel. Invited partners need to follow the invite_link_uri provided in the response to accept. After accepting the invitation, a link is set up between the two parties. You must be a distributor to call this method. Possible error codes: * PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request. * INVALID_ARGUMENT: Required request parameters are missing or invalid. * ALREADY_EXISTS: The ChannelPartnerLink sent in the request already exists. * NOT_FOUND: No Cloud Identity customer exists for provided domain. * INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. * UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. Return value: The new ChannelPartnerLink resource.

Arguments

request - No description provided.

parent - Required. Create a channel partner link for the provided reseller account's resource name. Parent uses the format: accounts/{account_id}

pub fn channel_partner_links_get(
&self,
name: &str
) -> AccountChannelPartnerLinkGetCall<'a, C>

Create a builder to help you perform the following task:

Returns a requested ChannelPartnerLink resource. You must be a distributor to call this method. Possible error codes: * PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request. * INVALID_ARGUMENT: Required request parameters are missing or invalid. * NOT_FOUND: ChannelPartnerLink resource not found because of an invalid channel partner link name. Return value: The ChannelPartnerLink resource.

Arguments

name - Required. The resource name of the channel partner link to retrieve. Name uses the format: accounts/{account_id}/channelPartnerLinks/{id} where {id} is the Cloud Identity ID of the partner.

pub fn channel_partner_links_list(
&self,
parent: &str
) -> AccountChannelPartnerLinkListCall<'a, C>

Create a builder to help you perform the following task:

List ChannelPartnerLinks belonging to a distributor. You must be a distributor to call this method. Possible error codes: * PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request. * INVALID_ARGUMENT: Required request parameters are missing or invalid. Return value: The list of the distributor account's ChannelPartnerLink resources.

Arguments

parent - Required. The resource name of the reseller account for listing channel partner links. Parent uses the format: accounts/{account_id}

pub fn channel_partner_links_patch(
&self,
request: GoogleCloudChannelV1UpdateChannelPartnerLinkRequest,
name: &str
) -> AccountChannelPartnerLinkPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a channel partner link. Distributors call this method to change a link's status. For example, to suspend a partner link. You must be a distributor to call this method. Possible error codes: * PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request. * INVALID_ARGUMENT: * Required request parameters are missing or invalid. * Link state cannot change from invited to active or suspended. * Cannot send reseller_cloud_identity_id, invite_url, or name in update mask. * NOT_FOUND: ChannelPartnerLink resource not found. * INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. * UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. Return value: The updated ChannelPartnerLink resource.

Arguments

request - No description provided.

name - Required. The resource name of the channel partner link to cancel. Name uses the format: accounts/{account_id}/channelPartnerLinks/{id} where {id} is the Cloud Identity ID of the partner.

pub fn customers_entitlements_activate(
&self,
request: GoogleCloudChannelV1ActivateEntitlementRequest,
name: &str
) -> AccountCustomerEntitlementActivateCall<'a, C>

Create a builder to help you perform the following task:

Activates a previously suspended entitlement. Entitlements suspended for pending ToS acceptance can't be activated using this method. An entitlement activation is a long-running operation and it updates the state of the customer entitlement. Possible error codes: * PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request. * INVALID_ARGUMENT: Required request parameters are missing or invalid. * NOT_FOUND: Entitlement resource not found. * SUSPENSION_NOT_RESELLER_INITIATED: Can only activate reseller-initiated suspensions and entitlements that have accepted the TOS. * NOT_SUSPENDED: Can only activate suspended entitlements not in an ACTIVE state. * INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. * UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. Return value: The ID of a long-running operation. To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of OperationMetadata.

Arguments

request - No description provided.

name - Required. The resource name of the entitlement to activate. Name uses the format: accounts/{account_id}/customers/{customer_id}/entitlements/{entitlement_id}

pub fn customers_entitlements_cancel(
&self,
request: GoogleCloudChannelV1CancelEntitlementRequest,
name: &str
) -> AccountCustomerEntitlementCancelCall<'a, C>

Create a builder to help you perform the following task:

Cancels a previously fulfilled entitlement. An entitlement cancellation is a long-running operation. Possible error codes: * PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request. * FAILED_PRECONDITION: There are Google Cloud projects linked to the Google Cloud entitlement's Cloud Billing subaccount. * INVALID_ARGUMENT: Required request parameters are missing or invalid. * NOT_FOUND: Entitlement resource not found. * DELETION_TYPE_NOT_ALLOWED: Cancel is only allowed for Google Workspace add-ons, or entitlements for Google Cloud's development platform. * INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. * UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. Return value: The ID of a long-running operation. To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The response will contain google.protobuf.Empty on success. The Operation metadata will contain an instance of OperationMetadata.

Arguments

request - No description provided.

name - Required. The resource name of the entitlement to cancel. Name uses the format: accounts/{account_id}/customers/{customer_id}/entitlements/{entitlement_id}

pub fn customers_entitlements_change_offer(
&self,
request: GoogleCloudChannelV1ChangeOfferRequest,
name: &str
) -> AccountCustomerEntitlementChangeOfferCall<'a, C>

Create a builder to help you perform the following task:

Updates the Offer for an existing customer entitlement. An entitlement update is a long-running operation and it updates the entitlement as a result of fulfillment. Possible error codes: * PERMISSION_DENIED: The customer doesn't belong to the reseller. * INVALID_ARGUMENT: Required request parameters are missing or invalid. * NOT_FOUND: Offer or Entitlement resource not found. * INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. * UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. Return value: The ID of a long-running operation. To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of OperationMetadata.

Arguments

request - No description provided.

name - Required. The resource name of the entitlement to update. Name uses the format: accounts/{account_id}/customers/{customer_id}/entitlements/{entitlement_id}

pub fn customers_entitlements_change_parameters(
&self,
request: GoogleCloudChannelV1ChangeParametersRequest,
name: &str
) -> AccountCustomerEntitlementChangeParameterCall<'a, C>

Create a builder to help you perform the following task:

Change parameters of the entitlement. An entitlement update is a long-running operation and it updates the entitlement as a result of fulfillment. Possible error codes: * PERMISSION_DENIED: The customer doesn't belong to the reseller. * INVALID_ARGUMENT: Required request parameters are missing or invalid. For example, the number of seats being changed is greater than the allowed number of max seats, or decreasing seats for a commitment based plan. * NOT_FOUND: Entitlement resource not found. * INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. * UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. Return value: The ID of a long-running operation. To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of OperationMetadata.

Arguments

request - No description provided.

name - Required. The name of the entitlement to update. Name uses the format: accounts/{account_id}/customers/{customer_id}/entitlements/{entitlement_id}

pub fn customers_entitlements_change_renewal_settings(
&self,
request: GoogleCloudChannelV1ChangeRenewalSettingsRequest,
name: &str
) -> AccountCustomerEntitlementChangeRenewalSettingCall<'a, C>

Create a builder to help you perform the following task:

Updates the renewal settings for an existing customer entitlement. An entitlement update is a long-running operation and it updates the entitlement as a result of fulfillment. Possible error codes: * PERMISSION_DENIED: The customer doesn't belong to the reseller. * INVALID_ARGUMENT: Required request parameters are missing or invalid. * NOT_FOUND: Entitlement resource not found. * NOT_COMMITMENT_PLAN: Renewal Settings are only applicable for a commitment plan. Can't enable or disable renewals for non-commitment plans. * INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. * UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. Return value: The ID of a long-running operation. To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of OperationMetadata.

Arguments

request - No description provided.

name - Required. The name of the entitlement to update. Name uses the format: accounts/{account_id}/customers/{customer_id}/entitlements/{entitlement_id}

pub fn customers_entitlements_create(
&self,
request: GoogleCloudChannelV1CreateEntitlementRequest,
parent: &str
) -> AccountCustomerEntitlementCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates an entitlement for a customer. Possible error codes: * PERMISSION_DENIED: The customer doesn't belong to the reseller. * INVALID_ARGUMENT: * Required request parameters are missing or invalid. * There is already a customer entitlement for a SKU from the same product family. * INVALID_VALUE: Make sure the OfferId is valid. If it is, contact Google Channel support for further troubleshooting. * NOT_FOUND: The customer or offer resource was not found. * ALREADY_EXISTS: * The SKU was already purchased for the customer. * The customer's primary email already exists. Retry after changing the customer's primary contact email. * CONDITION_NOT_MET or FAILED_PRECONDITION: * The domain required for purchasing a SKU has not been verified. * A pre-requisite SKU required to purchase an Add-On SKU is missing. For example, Google Workspace Business Starter is required to purchase Vault or Drive. * (Developer accounts only) Reseller and resold domain must meet the following naming requirements: * Domain names must start with goog-test. * Domain names must include the reseller domain. * INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. * UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. Return value: The ID of a long-running operation. To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of OperationMetadata.

Arguments

request - No description provided.

parent - Required. The resource name of the reseller's customer account in which to create the entitlement. Parent uses the format: accounts/{account_id}/customers/{customer_id}

pub fn customers_entitlements_get(
&self,
name: &str
) -> AccountCustomerEntitlementGetCall<'a, C>

Create a builder to help you perform the following task:

Returns a requested Entitlement resource. Possible error codes: * PERMISSION_DENIED: The customer doesn't belong to the reseller. * INVALID_ARGUMENT: Required request parameters are missing or invalid. * NOT_FOUND: The customer entitlement was not found. Return value: The requested Entitlement resource.

Arguments

name - Required. The resource name of the entitlement to retrieve. Name uses the format: accounts/{account_id}/customers/{customer_id}/entitlements/{entitlement_id}

pub fn customers_entitlements_list(
&self,
parent: &str
) -> AccountCustomerEntitlementListCall<'a, C>

Create a builder to help you perform the following task:

Lists Entitlements belonging to a customer. Possible error codes: * PERMISSION_DENIED: The customer doesn't belong to the reseller. * INVALID_ARGUMENT: Required request parameters are missing or invalid. Return value: A list of the customer's Entitlements.

Arguments

parent - Required. The resource name of the reseller's customer account to list entitlements for. Parent uses the format: accounts/{account_id}/customers/{customer_id}

pub fn customers_entitlements_start_paid_service(
&self,
request: GoogleCloudChannelV1StartPaidServiceRequest,
name: &str
) -> AccountCustomerEntitlementStartPaidServiceCall<'a, C>

Create a builder to help you perform the following task:

Starts paid service for a trial entitlement. Starts paid service for a trial entitlement immediately. This method is only applicable if a plan is set up for a trial entitlement but has some trial days remaining. Possible error codes: * PERMISSION_DENIED: The customer doesn't belong to the reseller. * INVALID_ARGUMENT: Required request parameters are missing or invalid. * NOT_FOUND: Entitlement resource not found. * FAILED_PRECONDITION/NOT_IN_TRIAL: This method only works for entitlement on trial plans. * INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. * UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. Return value: The ID of a long-running operation. To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of OperationMetadata.

Arguments

request - No description provided.

name - Required. The name of the entitlement to start a paid service for. Name uses the format: accounts/{account_id}/customers/{customer_id}/entitlements/{entitlement_id}

pub fn customers_entitlements_suspend(
&self,
request: GoogleCloudChannelV1SuspendEntitlementRequest,
name: &str
) -> AccountCustomerEntitlementSuspendCall<'a, C>

Create a builder to help you perform the following task:

Suspends a previously fulfilled entitlement. An entitlement suspension is a long-running operation. Possible error codes: * PERMISSION_DENIED: The customer doesn't belong to the reseller. * INVALID_ARGUMENT: Required request parameters are missing or invalid. * NOT_FOUND: Entitlement resource not found. * NOT_ACTIVE: Entitlement is not active. * INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. * UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. Return value: The ID of a long-running operation. To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of OperationMetadata.

Arguments

request - No description provided.

name - Required. The resource name of the entitlement to suspend. Name uses the format: accounts/{account_id}/customers/{customer_id}/entitlements/{entitlement_id}

pub fn customers_create(
&self,
request: GoogleCloudChannelV1Customer,
parent: &str
) -> AccountCustomerCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new Customer resource under the reseller or distributor account. Possible error codes: * PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request. * INVALID_ARGUMENT: * Required request parameters are missing or invalid. * Domain field value doesn't match the primary email domain. Return value: The newly created Customer resource.

Arguments

request - No description provided.

parent - Required. The resource name of reseller account in which to create the customer. Parent uses the format: accounts/{account_id}

pub fn customers_delete(&self, name: &str) -> AccountCustomerDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes the given Customer permanently and irreversibly. Possible error codes: * PERMISSION_DENIED: The account making the request does not own this customer. * INVALID_ARGUMENT: Required request parameters are missing or invalid. * FAILED_PRECONDITION: The customer has existing entitlements. * NOT_FOUND: No Customer resource found for the name in the request.

Arguments

name - Required. The resource name of the customer to delete.

pub fn customers_get(&self, name: &str) -> AccountCustomerGetCall<'a, C>

Create a builder to help you perform the following task:

Returns a requested Customer resource. Possible error codes: * PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request. * INVALID_ARGUMENT: Required request parameters are missing or invalid. * NOT_FOUND: The customer resource doesn't exist. Usually the result of an invalid name parameter. Return value: The Customer resource.

Arguments

name - Required. The resource name of the customer to retrieve. Name uses the format: accounts/{account_id}/customers/{customer_id}

pub fn customers_list(&self, parent: &str) -> AccountCustomerListCall<'a, C>

Create a builder to help you perform the following task:

List Customers. Possible error codes: * PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request. * INVALID_ARGUMENT: Required request parameters are missing or invalid. Return value: List of Customers, or an empty list if there are no customers.

Arguments

parent - Required. The resource name of the reseller account to list customers from. Parent uses the format: accounts/{account_id}.

pub fn customers_list_purchasable_offers(
&self,
customer: &str
) -> AccountCustomerListPurchasableOfferCall<'a, C>

Create a builder to help you perform the following task:

Lists the following: * Offers that you can purchase for a customer. * Offers that you can change for an entitlement. Possible error codes: * PERMISSION_DENIED: The customer doesn't belong to the reseller * INVALID_ARGUMENT: Required request parameters are missing or invalid.

Arguments

customer - Required. The resource name of the customer to list Offers for. Format: accounts/{account_id}/customers/{customer_id}.

pub fn customers_list_purchasable_skus(
&self,
customer: &str
) -> AccountCustomerListPurchasableSkuCall<'a, C>

Create a builder to help you perform the following task:

Lists the following: * SKUs that you can purchase for a customer * SKUs that you can upgrade or downgrade for an entitlement. Possible error codes: * PERMISSION_DENIED: The customer doesn't belong to the reseller. * INVALID_ARGUMENT: Required request parameters are missing or invalid.

Arguments

customer - Required. The resource name of the customer to list SKUs for. Format: accounts/{account_id}/customers/{customer_id}.

pub fn customers_patch(
&self,
request: GoogleCloudChannelV1Customer,
name: &str
) -> AccountCustomerPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates an existing Customer resource for the reseller or distributor. Possible error codes: * PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request. * INVALID_ARGUMENT: Required request parameters are missing or invalid. * NOT_FOUND: No Customer resource found for the name in the request. Return value: The updated Customer resource.

Arguments

request - No description provided.

name - Output only. Resource name of the customer. Format: accounts/{account_id}/customers/{customer_id}

pub fn customers_provision_cloud_identity(
&self,
request: GoogleCloudChannelV1ProvisionCloudIdentityRequest,
customer: &str
) -> AccountCustomerProvisionCloudIdentityCall<'a, C>

Create a builder to help you perform the following task:

Creates a Cloud Identity for the given customer using the customer's information, or the information provided here. Possible error codes: * PERMISSION_DENIED: The customer doesn't belong to the reseller. * INVALID_ARGUMENT: Required request parameters are missing or invalid. * NOT_FOUND: The customer was not found. * ALREADY_EXISTS: The customer's primary email already exists. Retry after changing the customer's primary contact email. * INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. * UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. Return value: The ID of a long-running operation. To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata contains an instance of OperationMetadata.

Arguments

request - No description provided.

customer - Required. Resource name of the customer. Format: accounts/{account_id}/customers/{customer_id}

pub fn customers_transfer_entitlements(
&self,
request: GoogleCloudChannelV1TransferEntitlementsRequest,
parent: &str
) -> AccountCustomerTransferEntitlementCall<'a, C>

Create a builder to help you perform the following task:

Transfers customer entitlements to new reseller. Possible error codes: * PERMISSION_DENIED: The customer doesn't belong to the reseller. * INVALID_ARGUMENT: Required request parameters are missing or invalid. * NOT_FOUND: The customer or offer resource was not found. * ALREADY_EXISTS: The SKU was already transferred for the customer. * CONDITION_NOT_MET or FAILED_PRECONDITION: * The SKU requires domain verification to transfer, but the domain is not verified. * An Add-On SKU (example, Vault or Drive) is missing the pre-requisite SKU (example, G Suite Basic). * (Developer accounts only) Reseller and resold domain must meet the following naming requirements: * Domain names must start with goog-test. * Domain names must include the reseller domain. * Specify all transferring entitlements. * INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. * UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. Return value: The ID of a long-running operation. To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of OperationMetadata.

Arguments

request - No description provided.

parent - Required. The resource name of the reseller's customer account that will receive transferred entitlements. Parent uses the format: accounts/{account_id}/customers/{customer_id}

pub fn customers_transfer_entitlements_to_google(
&self,
request: GoogleCloudChannelV1TransferEntitlementsToGoogleRequest,
parent: &str
) -> AccountCustomerTransferEntitlementsToGoogleCall<'a, C>

Create a builder to help you perform the following task:

Transfers customer entitlements from their current reseller to Google. Possible error codes: * PERMISSION_DENIED: The customer doesn't belong to the reseller. * INVALID_ARGUMENT: Required request parameters are missing or invalid. * NOT_FOUND: The customer or offer resource was not found. * ALREADY_EXISTS: The SKU was already transferred for the customer. * CONDITION_NOT_MET or FAILED_PRECONDITION: * The SKU requires domain verification to transfer, but the domain is not verified. * An Add-On SKU (example, Vault or Drive) is missing the pre-requisite SKU (example, G Suite Basic). * (Developer accounts only) Reseller and resold domain must meet the following naming requirements: * Domain names must start with goog-test. * Domain names must include the reseller domain. * INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. * UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. Return value: The ID of a long-running operation. To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The response will contain google.protobuf.Empty on success. The Operation metadata will contain an instance of OperationMetadata.

Arguments

request - No description provided.

parent - Required. The resource name of the reseller's customer account where the entitlements transfer from. Parent uses the format: accounts/{account_id}/customers/{customer_id}

pub fn offers_list(&self, parent: &str) -> AccountOfferListCall<'a, C>

Create a builder to help you perform the following task:

Lists the Offers the reseller can sell. Possible error codes: * INVALID_ARGUMENT: Required request parameters are missing or invalid.

Arguments

parent - Required. The resource name of the reseller account from which to list Offers. Parent uses the format: accounts/{account_id}.

pub fn check_cloud_identity_accounts_exist(
&self,
request: GoogleCloudChannelV1CheckCloudIdentityAccountsExistRequest,
parent: &str
) -> AccountCheckCloudIdentityAccountsExistCall<'a, C>

Create a builder to help you perform the following task:

Confirms the existence of Cloud Identity accounts based on the domain and if the Cloud Identity accounts are owned by the reseller. Possible error codes: * PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request. * INVALID_ARGUMENT: Required request parameters are missing or invalid. * INVALID_VALUE: Invalid domain value in the request. Return value: A list of CloudIdentityCustomerAccount resources for the domain (may be empty) Note: in the v1alpha1 version of the API, a NOT_FOUND error returns if no CloudIdentityCustomerAccount resources match the domain.

Arguments

request - No description provided.

parent - Required. The reseller account's resource name. Parent uses the format: accounts/{account_id}

pub fn list_subscribers(
&self,
account: &str
) -> AccountListSubscriberCall<'a, C>

Create a builder to help you perform the following task:

Lists service accounts with subscriber privileges on the Cloud Pub/Sub topic created for this Channel Services account. Possible error codes: * PERMISSION_DENIED: The reseller account making the request and the provided reseller account are different, or the impersonated user is not a super admin. * INVALID_ARGUMENT: Required request parameters are missing or invalid. * NOT_FOUND: The topic resource doesn't exist. * INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. * UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. Return value: A list of service email addresses.

Arguments

account - Required. Resource name of the account.

pub fn list_transferable_offers(
&self,
request: GoogleCloudChannelV1ListTransferableOffersRequest,
parent: &str
) -> AccountListTransferableOfferCall<'a, C>

Create a builder to help you perform the following task:

List TransferableOffers of a customer based on Cloud Identity ID or Customer Name in the request. Use this method when a reseller gets the entitlement information of an unowned customer. The reseller should provide the customer's Cloud Identity ID or Customer Name. Possible error codes: * PERMISSION_DENIED: * The customer doesn't belong to the reseller and has no auth token. * The supplied auth token is invalid. * The reseller account making the request is different from the reseller account in the query. * INVALID_ARGUMENT: Required request parameters are missing or invalid. Return value: List of TransferableOffer for the given customer and SKU.

Arguments

request - No description provided.

parent - Required. The resource name of the reseller's account.

pub fn list_transferable_skus(
&self,
request: GoogleCloudChannelV1ListTransferableSkusRequest,
parent: &str
) -> AccountListTransferableSkuCall<'a, C>

Create a builder to help you perform the following task:

List TransferableSkus of a customer based on the Cloud Identity ID or Customer Name in the request. Use this method to list the entitlements information of an unowned customer. You should provide the customer's Cloud Identity ID or Customer Name. Possible error codes: * PERMISSION_DENIED: * The customer doesn't belong to the reseller and has no auth token. * The supplied auth token is invalid. * The reseller account making the request is different from the reseller account in the query. * INVALID_ARGUMENT: Required request parameters are missing or invalid. Return value: A list of the customer's TransferableSku.

Arguments

request - No description provided.

parent - Required. The reseller account's resource name. Parent uses the format: accounts/{account_id}

pub fn register(
&self,
request: GoogleCloudChannelV1RegisterSubscriberRequest,
account: &str
) -> AccountRegisterCall<'a, C>

Create a builder to help you perform the following task:

Registers a service account with subscriber privileges on the Cloud Pub/Sub topic for this Channel Services account. After you create a subscriber, you get the events through SubscriberEvent Possible error codes: * PERMISSION_DENIED: The reseller account making the request and the provided reseller account are different, or the impersonated user is not a super admin. * INVALID_ARGUMENT: Required request parameters are missing or invalid. * INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. * UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. Return value: The topic name with the registered service email address.

Arguments

request - No description provided.

account - Required. Resource name of the account.

pub fn unregister(
&self,
request: GoogleCloudChannelV1UnregisterSubscriberRequest,
account: &str
) -> AccountUnregisterCall<'a, C>

Create a builder to help you perform the following task:

Unregisters a service account with subscriber privileges on the Cloud Pub/Sub topic created for this Channel Services account. If there are no service accounts left with subscriber privileges, this deletes the topic. You can call ListSubscribers to check for these accounts. Possible error codes: * PERMISSION_DENIED: The reseller account making the request and the provided reseller account are different, or the impersonated user is not a super admin. * INVALID_ARGUMENT: Required request parameters are missing or invalid. * NOT_FOUND: The topic resource doesn't exist. * INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. * UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support. Return value: The topic name that unregistered the service email address. Returns a success response if the service email address wasn't registered with the topic.

Arguments

request - No description provided.

account - Required. Resource name of the account.